% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_test.R
\name{match_test}
\alias{match_test}
\title{Verify that lottery output meets constraints imposed by source data}
\usage{
match_test(
  dir_review,
  match,
  gradelevels,
  contactsmatch,
  choices,
  eps_gradelevel,
  eps_choice
)
}
\arguments{
\item{dir_review}{character}

\item{match}{tibble of choice-level match outcomes}

\item{gradelevels}{tibble of grade level records}

\item{contactsmatch}{tibble of contact records}

\item{choices}{tibble of application school ranking records}

\item{eps_gradelevel}{tibble of eligibility and priority information
for all grade levels in the lottery}

\item{eps_choice}{tibble of eligibility and priority information
for all application school rankings in the lottery}
}
\description{
Verify that lottery output meets constraints imposed by source data
}
